Remove python-libvirt autodep on clara-core pkg
dh_python2 guesses that clara-core depends on python-libvirt because
the requires.txt generated by setuptools in egg tree tells it is a dep
of clara codebase. This is absolutely true but the dep is only required
by the code finally installed in clara-plugin-virt plugin. This is why
this package explicitely depends on this library.

To avoid python-libvirt being installed when only installing clara-core
package, this commit overrides pydist to remove the automatic
dependency on this library.